Renamed staging_lookup -> tnmstage_lookup.
