+++ test/backend/lib/staging.test.js
@@ -1,6 +1,6 @@
 import {expect} from 'chai';
 import staging from '../../../src/lib/staging.jsx';
-import lookup from '../../../src/lib/staging_lookup.jsx';
+import lookup from '../../../src/lib/tnmstage_lookup.jsx';
 
 //Helpers
 function metastaticPermute(ts, ns) {
